refactor/ change folder of CopyList
