+++ src/components/CopyList/style.ts
@@ -1,0 +1,15 @@
+import { css } from "lit";
+
+export default css`
+  :host {
+    max-width: 1280px;
+    margin: 0 auto;
+    padding: 2rem;
+    text-align: center;
+  }
+
+  ul {
+    list-style: none;
+    padding: none;
+  }
+`;
